+++ src/main/java/com/mysite/sbb/question/QuestionRepository.java
@@ -1,4 +1,4 @@
-package com.mysite.sbb;
+package com.mysite.sbb.question;
 
 import org.springframework.data.jpa.repository.JpaRepository;
